+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Fixed
 - Inputs and configurations for tools that were not opened were lost the next time the settings were saved.
 - Loading the tool window settings causes an error if one of the tools has a saved secret.
+- In the JWT Encoder/Decoder the "Public Key" field was visible for the HMAC signature algorithm.
 
 ## 3.2.0 - 2024-01-06